+++ src/char/int_guild.h
@@ -12,7 +12,7 @@ void inter_guild_final(void);
 int inter_guild_parse_frommap(int fd);
 void inter_guild_mapif_init(int fd);
 void inter_guild_leave(int guild_id, int account_id, int char_id);
-void inter_guild_sex_changed(int guild_id,int account_id,int char_id, int gender);
+void inter_guild_sex_changed(int guild_id,int account_id,int char_id, short gender);
 void inter_guild_charname_changed(int guild_id, int account_id, int char_id, const char* name);
 void mapif_parse_BreakGuild(int fd, int guild_id);
